Extract and fix change_admin and set_contract_status tests
